Rename cisco_system model
